Move participant list implementation to according directory
